Moved InterviewSummaryComponent
