--- app/components/candidate_interface/continuous_applications/interview_summary_component.rb
@@ -1,25 +1,0 @@
-module CandidateInterface
-  module ContinuousApplications
-    class InterviewSummaryComponent < ViewComponent::Base
-      attr_reader :interview
-
-      def initialize(interview:)
-        @interview = interview
-      end
-
-      delegate :location, to: :interview
-
-      def time
-        interview.date_and_time.to_fs(:govuk_time)
-      end
-
-      def date
-        interview.date_and_time.to_fs(:govuk_date)
-      end
-
-      def details
-        interview.additional_details
-      end
-    end
-  end
-end
