+++ app/components/candidate_interface/interview_summary_component.html.erb
