+++ app/components/candidate_interface/interview_summary_component.rb
@@ -1,0 +1,23 @@
+module CandidateInterface
+  class InterviewSummaryComponent < ViewComponent::Base
+    attr_reader :interview
+
+    def initialize(interview:)
+      @interview = interview
+    end
+
+    delegate :location, to: :interview
+
+    def time
+      interview.date_and_time.to_fs(:govuk_time)
+    end
+
+    def date
+      interview.date_and_time.to_fs(:govuk_date)
+    end
+
+    def details
+      interview.additional_details
+    end
+  end
+end
